requirements.txt
python-dotenv~=1.1
pydantic>=2.6,<3
langchain==0.3.*
langchain-core==0.3.*
langchain-community==0.3.*
langchain-chroma==0.1.2
chromadb==0.5.11
redis==6.4.0
requests~=2.32.4
langgraph~=0.6.5